Implement payloads for state transitions
Now, entry handlers (may) take an argument holding a payload passed to the
runtime through `gotoState()`.

There are a few sharp corners here which I don't like: in particular, because
we can't guarantee that all payloads will be of the same type, we have to defer
to runtime type checking - in other words, an ill-typed entry handler could
throw an exception.  Additionally, because not every goto takes a payload,
currently the argument can potentially be null.  I might change this depending
on what folks think - it could take an Optional<T> rather than a T; it's more
verbose but safer.